requirements.txt
asgiref==3.8.1
astroid==2.13.5
colorama==0.4.6
dill==0.3.9
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
isort==5.13.2
lazy-object-proxy==1.10.0
mccabe==0.7.0
pip-chill==1.0.3
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyJWT==2.10.1
pylint==2.15.10
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
python-decouple==3.8
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
wrapt==1.17.0
